Available bike store wise

+++ BikeBookingApplicationVegoBike/src/main/java/com/eptiq/vegobike/dtos/AvailableBikeSearchRequest.java
@@ -10,7 +10,7 @@ import java.util.Date;
 @NoArgsConstructor
 public class AvailableBikeSearchRequest {
     private Integer cityId;       // City where the booking is needed
-    private Integer storeId;      // Store to filter available bikes
+    private Long storeId;      // Store to filter available bikes
     private Date startDate;       // Desired booking/pickup date and time
     private Date endDate;         // Desired drop-off date and time
     private String searchText;    // (Optional) Search string for bike model, registration, etc.
